finished imgui support!

+++ src/gg/camera.c
@@ -1,6 +1,6 @@
 #include "camera.h"
 
-void Camera_Create(gg_camera_t* camera, uint32_t window_w, uint32_t window_h, uint32_t x, uint32_t y) {
+void Camera_Create(gg_camera_t* camera, uint32_t window_w, uint32_t window_h, float x, float y) {
     camera->x = x;
     camera->y = y;
     camera->rotation = 0.f;
