+++ src/gg/camera.h
@@ -10,6 +10,6 @@ typedef struct gg_camera {
     float rotation;
 } gg_camera_t;
 
-void Camera_Create(gg_camera_t* camera, uint32_t window_w, uint32_t window_h, uint32_t x, uint32_t y);
+void Camera_Create(gg_camera_t* camera, uint32_t window_w, uint32_t window_h, float x, float y);
 
 #endif  // GG_CAMERA_H
